+++ src/components/Toast/index.e2e.tsx
@@ -1,0 +1,5 @@
+export function ToastContainer() {
+  return null
+}
+
+export function show() {}
