+++ src/components/Toast/types.ts
@@ -1,0 +1,1 @@
+export type ToastType = 'default' | 'success' | 'error' | 'warning' | 'info'
